app/requirements.txt
altgraph==0.17.2
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
boto3==1.18.20
botocore==1.21.20
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
Django==3.2.6
django-filter==2.4.0
django-resized==0.3.11
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
fpdf==1.7.2
future==0.18.2
gunicorn==20.1.0
idna==3.2
img2pdf==0.4.1
jmespath==0.10.0
lxml==4.6.3
numpy==1.21.0
openpyxl==3.0.9
pandas==1.3.0
pathlib==1.0.1
pdfkit==0.6.1
pdfrw==0.4
pefile==2021.9.3
pikepdf==3.1.0
Pillow==8.3.1
platformdirs==2.2.0
progressbar==2.5
psutil==5.8.0
psycopg2==2.9.1
pycodestyle==2.7.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyMsgBox==1.0.9
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==301;platform_system=="Windows"
pywin32-ctypes==0.2.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
ttkthemes==3.2.2
urllib3==1.26.6
virtualenv==20.6.0
whitenoise==5.3.0
xlwings==0.24.9;platform_system=="Windows"
